backend/requirements.txt
fastapi==0.128.0
pydantic[email]==2.12.5
uvicorn[standard]==0.40.0
python-multipart==0.0.22
sqlalchemy==2.0.45
alembic==1.18.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
sendgrid==6.12.5
httpx==0.28.1